added warren's cdriz_setup, lanczos3 still fails.
