+++ tests/drizzle/test_cdriz.py
@@ -1,12 +1,12 @@
 import pytest
 import numpy as np
-import test_setup
+import cdriz_setup
 
 
 @pytest.fixture
 def kernel_pars():
-    # get cdriz.triz inputs from test_setup.py
-    _params = test_setup.Get_Grid(inx=3, iny=3, outx=4, outy=4)
+    # get cdriz.triz inputs from cdriz_setup.py
+    _params = cdriz_setup.Get_Grid(inx=5, iny=5, outx=10, outy=10)
     _params.zero_background()
     return _params
 
@@ -26,31 +26,37 @@ def test_square_kernel(kernel_pars, kernel, return_png=True):
         Whether to return a png of the outputs.
     """
     # add bright pixel at center
-    kernel_pars.insci[1:2, 1:2] = 1e4
+    kernel_pars.insci[3, 3] = 1e4
 
     # resample:
-    test_setup.cdriz_call(kernel_pars, kernel)
+    cdriz_setup.cdriz_call(kernel_pars, kernel)
 
     if return_png:
         # save truth file as figure
-        test_setup.generate_png(kernel_pars, "cdriz_square.png")
+        cdriz_setup.generate_png(
+            kernel_pars, "./tests/drizzle/truth_files/cdriz_square.png"
+        )
 
-    assert np.allclose(np.sum(kernel_pars.outsci), 5000, 1e-7)
+    assert np.allclose(np.sum(kernel_pars.outsci), 10000, 1e-7)
 
 
 def test_gaussian_kernel(kernel_pars, return_png=True):
     """Same as above test but with gaussian kernel."""
-    kernel_pars.insci[1:2, 1:2] = 1e4
-    test_setup.cdriz_call(kernel_pars, "gaussian")
+    kernel_pars.insci[3, 3] = 1e4
+    cdriz_setup.cdriz_call(kernel_pars, "gaussian")
     if return_png:
-        test_setup.generate_png(kernel_pars, "cdriz_gaussian.png")
-    assert np.allclose(np.sum(kernel_pars.outsci), 5314.6846, 1e-3)
+        cdriz_setup.generate_png(
+            kernel_pars, "./tests/drizzle/truth_files/cdriz_gaussian.png"
+        )
+    assert np.allclose(np.sum(kernel_pars.outsci), 10058.42, 1e-3)
 
 
 def test_lanczos3_kernel(kernel_pars, return_png=True):
     """Same as above test but with lanczos3 kernel."""
-    kernel_pars.insci[1:2, 1:2] = 1e4
-    test_setup.cdriz_call(kernel_pars, "lanczos3")
+    kernel_pars.insci[3, 3] = 1e4
+    cdriz_setup.cdriz_call(kernel_pars, "lanczos3")
     if return_png:
-        test_setup.generate_png(kernel_pars, "cdriz_lanczos3.png")
-    assert np.allclose(np.sum(kernel_pars.outsci), 4960.454, 1e-3)
+        cdriz_setup.generate_png(
+            kernel_pars, "./tests/drizzle/truth_files/cdriz_lanczos3.png"
+        )
+    assert np.allclose(np.sum(kernel_pars.outsci), 1E4, 1e-3)
